requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
groq==0.4.1
pydantic==2.5.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0